+++ mobile_app/android/app/src/main/kotlin/com/ankhoi/hcareplus/hcare_plus/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.ankhoi.hcareplus.hcare_plus
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()